requirements.txt
python=3.7
pandas
requests
google-cloud-bigquery